Merged export_all_partition_by_date.py into export_all.py
Changes are:
(1) <start_block> is changed to more generic <start>, which can
take either block number, date or Unix timestamp. Type is changed
to str and default value is removed (which goes well with the
original usage message).
(2) Some help functions are defined to check the argument type.
(3) The rest of the change is checking for argument types and
computing appropriate partitions for export_all()